Move FileFormatDataSourceStats to trino-plugin-toolkit module
