+++ lib/trino-plugin-toolkit/src/main/java/io/trino/plugin/base/metrics/FileFormatDataSourceStats.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.hive;
+package io.trino.plugin.base.metrics;
 
 import io.airlift.stats.DistributionStat;
 import io.airlift.stats.TimeStat;
