+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestOriginalFilesUtils.java
@@ -15,6 +15,7 @@ package io.trino.plugin.hive;
 
 import io.trino.filesystem.Location;
 import io.trino.orc.OrcReaderOptions;
+import io.trino.plugin.base.metrics.FileFormatDataSourceStats;
 import io.trino.plugin.hive.orc.OriginalFilesUtils;
 import org.junit.jupiter.api.Test;
