+++ plugin/trino-iceberg/src/main/java/io/trino/plugin/iceberg/TrinoOrcDataSource.java
@@ -19,7 +19,7 @@ import io.trino.filesystem.TrinoInputFile;
 import io.trino.orc.AbstractOrcDataSource;
 import io.trino.orc.OrcDataSourceId;
 import io.trino.orc.OrcReaderOptions;
-import io.trino.plugin.hive.FileFormatDataSourceStats;
+import io.trino.plugin.base.metrics.FileFormatDataSourceStats;
 
 import java.io.IOException;
